Github Update!
Even the Filterscript may not be updated anymore, I'd like to put it in
here anyway if someone wants to change something.

Changelog

Version 1: 24/08/2011 [dd/mm/yyyy] - First Release

Version 1.1:
- The countdown will print on your 'samp-server'. [Useful for who don't
want to open SAMP and refresh the server to see the countdown
everysecond. ]
- Fixed the countdown when goes to hour 0 it printed -1

Version 1.1 R1: Fixed the countdown when goes to hour 0 and minute 0 it
printed - sometime 0 or -1 to hour and 60 to minutes

Version 1.x Final
- I finnaly fixed all the issues about CountDown and printing.
- Now the countdown will be around 96% perfect in the count.
- When the countdown goes to hour 0, minute 0, seconds 0. [Will open the
server with your Original Server Name. Remember to set it in the FS]

Version 2
- Now the filterscript will be unloaded when the countdown finish
beacuse it's being useless after it. [Not completed yet cause of some
issues.]
- Now the CountDown will save every hour, minute, second thanks to DINI
by DragoneBlue. [ Very useful if you want to restart/make offline for
some seconds your GameMode without losing your countdown. When the
countdown will go to 0h 0m 0s, the FS will load the default time made in
the script. ]
- More precision in countdown it will be around 98-99% [Laste Update was
a bit late. On 1 hour of countdown it was about 1 minute late. ]
- Fixed the issue when you close the 'server-samp' at the same time as
0h 0m 0s it saved a negative random number. [Useful fix beacause if you
start the server and the file have a negative number will countdown it
instead the normai. ]
- Little thanks to Valve Secure beacause of save disk space [1 day =
about 50mb is not useful]. Now it's optional.

Version 2.0.1
- Converted the write/read file system (Thanks to Double-O-Seven)
Note: The functions still the same as dini beacuse i've just converted
into DOF2

Version 2.0.2
- Re-ordered some parts of the script to simply config your countdown.
- Many things re-written:
- Now it's more faster and more easy even newbies can take a look.
- Less lines used
- New Saving method

- Fixed the counting when starts from 0 minutes it went to -1
- Now the counting is precise at 99.9%
- Quick Config: - Just edit the "Config" lines to set your options.
- Saving & Loading the countdown file system changed to Y_Less's one.
- Days future added.

Version 2.0.2 R1
- Now you can lock/unlock your server until the countdown finishes
- You can enable/disable the password until the countdown finishes [Just
change "ServerPassword" variable to " true " or " false "]